test: ServiceTest Repository 의존성 부모 클래스로 이동
